Tests/LibWeb/Text/expected/css/getComputedStyle-relative-property-values.txt

height: 4px
width: 5px
